Move conversation-store, conversation-list-state and conversations-client from src/stores to src/stores/conversation
